… .travis.yml
… includes/class-woocommerce-custom-orders-table-filters.php
… tests/test-order-data-store.php
… tests/testcase.php