## Problem overview.
Even with the custom orders table some of the order fields are still stored in `postmeta` table and when queries relay in both the new orders table and `postmeta` are not much more faster.

## Problem details.
The plugin is migrating from `postmeta` to a new orders table the fields defined in [the mapping](https://github.com/liquidweb/woocommerce-custom-orders-table/blob/develop/includes/class-wc-order-data-store-custom-table.php#L135) which is basically the fields in the WC core Order object.

**But that Order object is commonly extended by other plugins. This common scenario is not covered by this plugin.** Payment gateways plugins like stripe or paywall, multilingual, EU-VAT plugins and others extensively used are extending the Order properties. In this case this plugin is not handling this fields to the new table, so are stored in `postmeta`.

In the described case most of the Order CRUD operations are not much faster because WooCommerce still queries both the orders table and `postmeta`.

## Proposed solution.
This PR implements a `ordersmeta` table. This table have the same structure of `postmeta` (`meta_id`, `order_id`, `meta_key` and `meta_value`) and stores or the meta information related to orders.

With this solution no more queries are required to `postmeta` table. It really enfast the Order operations.

Not that the PR passes all plugin tests and also all WooCommerce test. That means that also supports the legacy meta access such as `get_post_meta` and `update_post_meta`. Even this should never happen because plugins must use WooCommerce CRUD it takes care to check this access and prevent them. In this case it logs it to let developers fix his/her plugins to use the proper CRUD operations.

## Some real data about it.
In our use case are running a WooCommerce site with few plugins:
```
wc-duplicate-order
woocommerce
woocommerce-custom-orders-table
woocommerce-eu-vat-compliance-premium
woocommerce-auto-added-coupons
woocommerce-force-sells
woocommerce-order-status-manager
woocommerce-pdf-invoices-packing-slips
woocommerce-pdf-ips-pro
woocommerce-product-price-based-on-countries
woocommerce-price-based-country-pro-addon
woocommerce-shipment-tracking
woocommerce-gateway-stripe
woocommerce-subscriptions
woocommerce-subscriptions-cancel-on-refund
woocommerce-subscriptions-disable-report-cache-updates
woocommerce-subscriptions-importer-exporter
woo-orders-date-range-filter
woo-subscriptions-variation-lifetimeonetime     
```
Is a orders-intensive site with ~46k orders and this are our entries and times:
```
Without the plugin:
 - `postmeta` table rows: ~4M
 - Load order list: 2.7s
 - Open an order edit page: 30s
 - Save an order update: 30s

With the plugin without this PR:
 - `postmeta` table rows: ~3.8M
 - Load order list: 2.8s
 - Open an order edit page: 29s
 - Save an order update: 21s

With the plugin with this PR:
 - `postmeta` table rows: 600k
 - Load order list: 3s
 - Open an order edit page: 3.6s
 - Save an order update: 2.7
```
Ping me if you want any other specific date about it.

We would like to see this merged into the plugin at some point. Passing all tests looks solid enough to at least be tested. 

Please, test it and provide insights! Any feedback is welcome.