@juliquiron 2019-09-13
hi @tonybahama!
Unfortunatelly `get_post_meta()` hardcodes the post type (which make sense given the name):
```
function get_post_meta( $post_id, $key = '', $single = false ) {
    return get_metadata( 'post', $post_id, $key, $single );
}
```

https://developer.wordpress.org/reference/functions/get_post_meta/

Hopefully we can do: `get_metadata( 'order', $order_id, $key, $single );` and get the order meta. External plugins will need to be reviewed.

We can evaluate to implement the filter `get_post_metadata` to retrieve it and ensure that the `get_post_meta` outputs the correct one, but not done yet because with just `$object_id` do a select to check if is order can be a performance killer