@juliquiron 2019-09-17
Last commit adds support for the 'metabox' in the edit form. It allows to edit and add 'Custom Fields' to orders working with the new `woocommerce_ordermeta` table.

With this the CRUD operations are covered and only the plugins accessing metadata via `{action}_post_meta()` instead of the woocommerce CRUD will miss it.

The code that belongs to this metaboxes should be moved to a custom class, but the feature is already there and I pushed to let you test it and provide feedback.

FYI, in my case I moved from a `postmeta` table of 4M rows to a 600k one, with `woocommerce_ordermeta` of 2M and the `woocommerce_orders` table with 45k.

Hope it's useful for more people!