@juliquiron 2019-09-25
All tests passing.

At this point strongly pinging @stevegrunwell and @bswatson to know your feedback. 

I know form the WC slack discussions that you are working to see the orders table into WC core, but we need to care also about other metadata, not only the core fields, you see the times.

Thanks a lot for your work and your feedback.