@glagonikas 2019-10-14
Hey @stevegrunwell 

Our site is one that would certainly benefit from this PR and we're eager to see this merged.

I see your point about being backwards compatible however considering how long it's been since the introduction of the CRUD functions, most plugins should be using them and they should be forced to update anyway (unless they do very complex queries).

Considering this plugin requires WP CLI, the average user won't be installing v1.0.0 and the ones who would, can probably understand the possible workarounds.

One possible workaround would be to add a config to toggle the use of this PR until a decision is made on either breaking backwards compatibility or adding more customisations. Either way, the performance improvements of this PR look impressive and it's a shame to not merge it for v1.0.0.

Also, unless I'm mistaken, the PR doesn't take into account the meta of orders that have already moved?