@juliquiron 2019-10-31
Hi @stevegrunwell 

thanks for your comment and sorry for my late reply.

I will clarify the topics you mention in your comment:
 - The numbers are awesome **in our case**, but even will be relevant for other cases we cannot assume that the impact will be such big, for sure.
 - About backward compatibility: the actual code in this PR works even with plugins are using pre-CRUD operations such as `update_post_meta()`, as we can see in passing all the WC tests. So nothing will break. **But** , the way to achieve it is not so performant. Is introducing an extra query to figure out if the meta is from an order or not. The idea is to let pay in terms of performance the cases where the plugins are not updated, not the ones using the proper CRUD operations. See: https://github.com/liquidweb/woocommerce-custom-orders-table/pull/136/files#diff-e11dad352d57391a1b5028dfae99fa65R189

Anyway, the part handling the pre-CRUD operations can be worked to be turned on/off as @glagonikas suggest with a filter or a configuration option. 

about @glagonikas question: sure, the meta already managed in the order table is not taken into account for this PR. The code focus on the ones not migrated using the same mapping definition.

Thanks a lot for the feedback and feel free to ping me for anything I can do to move it forward.

BTW: nice talk in WooSesh ;)