@dcx15 2020-05-22
Hi guys, I just wanted to add my +1 for this. Thank you @juliquiron for the PR!

@stevegrunwell, I'm one of the (probably many) people which has been watching this project develop like an excited kid before Christmas. I am very much looking forward to 1.0.0 and can't thank you enough for your work on it!

However, missing out 3rd party meta does diminish the usefulness from my perspective. I think every WooCommerce site has a fairly large amount of 3rd party meta on orders, such as the core PayPal payment gateway which adds 4 fields per order:

- Payer first name
- Payer last name
- Payer PayPal address
- Payment type

Couple that with various shipping plugins and VAT data, the meta can get really out of hand fast.

I understand the reasons to look at adding this PR as a "phase 2" kinda thing. For a 1.0.0 release, is there a happy medium whereby people could mark specific meta keys as compatible via a filter? Then it is down to the developer's discretion which extra fields are processed via the custom table and which remain in postmeta, and normal users won't be impacted with the core meta as default.