@binaryfire 2020-06-01
Is there a way to add extra fields to the default mapping? (https://github.com/liquidweb/woocommerce-custom-orders-table/blob/develop/includes/class-wc-order-data-store-custom-table.php#L135).

I know modifying core tables normally isn't the greatest idea, but perhaps a filter could be added to allows devs to map additional fields to columns. As long as schema updates can take this into account, it should work...? This would be the most performant solution for custom development.