@lkraav 2020-06-01
> I know modifying core tables normally isn't the greatest idea, but perhaps a filter could be added to allows devs to map additional fields to columns. As long as schema updates can take this into account, it should work...? This would be the most performant solution for custom development.

Sounds sensible. Initial migration might be on the simple side, but how do you handle later schema changes - for example you later add a new plugin to the stack, and want it to start recording in custom tables as well.