@binaryfire 2020-06-13
> > Sounds sensible. Initial migration might be on the simple side, but how do you handle later schema changes - for example you later add a new plugin to the stack, and want it to start recording in custom tables as well.> > 

Wouldn't it be fairly simple to add a new column if it doesn't exist?

```
$row = $wpdb->get_results(  "SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.COLUMNS
WHERE table_name = 'wc_orders_table' AND column_name = 'new_column'"  );

if(empty($row)){
   $wpdb->query("ALTER TABLE wc_orders_table ADD new_column INT(1) NOT NULL DEFAULT 1");
}

```

My dev experience is limited so I may be missing something. But this is how I currently do it.